Implemented error control into insertion, updating and deletion endpoints

+++ src/controller.js
@@ -14,27 +14,50 @@ class BooksController{
     }
 
     async insert(req, res){
-        const book = req.body;
-        const [result] = await poolRequest.query(
-                `INSERT INTO books(name, author, category, publicationDate, isbn) VALUES(?, ?, ?, ?, ?)`, 
-                [book.name, book.author, book.category, book.publicationDate, book.isbn]
-            );
-        res.json({"Book added: ": result.insertId});
+        try{
+            const book = req.body;
+            const [result] = await poolRequest.query(
+                    `INSERT INTO books(name, author, category, publicationDate, isbn) VALUES(?, ?, ?, ?, ?)`, 
+                    [book.name, book.author, book.category, book.publicationDate, book.isbn]
+                );
+            res.json({"Book added: ": result.insertId});
+        }
+        catch(e){
+            res.json({"Error during insertion new Book => ": e});
+        }
     }
 
     async update(req, res){
-        const book = req.body;
-        const [result] = await poolRequest.query(
-            `UPDATE books SET name = ?, author = ?, category = ?, publicationDate = ?, isbn = ? WHERE id = ?`,
-            [book.name, book.author, book.category, book.publicationDate, book.isbn, book.id]
-        );
-        res.json({"Book updated" : result.affectedRows});
+        try{
+            const book = req.body;
+            if(book.id == null || book.name == null || book.author == null || book.publicationDate == null || book.isbn == null)
+                throw new Error("One or more attributes are invalids");
+
+            const [result] = await poolRequest.query(
+                `UPDATE books SET name = ?, author = ?, category = ?, publicationDate = ?, isbn = ? WHERE id = ?`,
+                [book.name, book.author, book.category, book.publicationDate, book.isbn, book.id]
+            );
+            if(result.affectedRows == 0) throw new Error("Book not found");
+            res.json({"Book updated" : result.affectedRows});
+        }
+        catch(e){
+            res.json({"ERROR ocurred during book Updating => ": e.message});
+        }
     }
 
     async deleteByIsbn(req, res){
-        const book = req.body;
-        const [result] = await poolRequest.query(`DELETE FROM books WHERE isbn = ?`, [book.isbn]);
-        res.json({"Book deleted" : result.affectedRows});
+        try{
+            const book = req.body;
+            if(book.isbn == null) throw Error("Invalid attribute"); //Forcing throw error according to some cases
+            const [result] = await poolRequest.query(`DELETE FROM books WHERE isbn = ?`, [book.isbn]);
+
+            if(result.affectedRows == 0) throw new Error("ISBN not found"); //Forcing throw error according to some cases
+
+            res.json({"Book deleted" : result.affectedRows});
+        }
+        catch(e){
+            res.json({"ERROR ocurred during book deletion => ": e.message});
+        }
     }
 }
